requirements.txt
ansible
requests
pyyaml
click
Flask
Jinja2
